+++ meraki_client/aio/_api/organizations.py
@@ -113,7 +113,6 @@ from meraki_client.schemas import (
     GetOrganizationInventoryDevicesResponseItem,
     GetOrganizationInventoryOnboardingCloudMonitoringImportsResponseItem,
     GetOrganizationLicensesOverviewResponse,
-    GetOrganizationNetworksResponseItem,
     GetOrganizationPoliciesAssignmentsByClientResponseItem,
     GetOrganizationPoliciesGlobalFirewallApplicationCategoriesResponseItem,
     GetOrganizationPoliciesGlobalGroupPoliciesAdaptivePolicyGroupsAssignmentsResponseItemsItem,
@@ -10010,7 +10009,7 @@ class Organizations:
         ending_before: str | None = None,
         total_pages: int | Literal["all"] = "all",
         direction: Literal["prev", "next"] = "next",
-    ) -> AsyncPaginatedResponse[GetOrganizationNetworksResponseItem]:
+    ) -> AsyncPaginatedResponse[CreateOrganizationNetworkResponse]:
         """List the networks that the user has privileges on in an organization.
 
         [API documentation: getOrganizationNetworks](https://developer.cisco.com/meraki/api-v1/#!get-organization-networks)
@@ -10105,7 +10104,7 @@ class Organizations:
             params=params,
             total_pages=total_pages,
             direction=direction,
-            item_schema=GetOrganizationNetworksResponseItem,
+            item_schema=CreateOrganizationNetworkResponse,
         )
 
     async def create_organization_network(